requirements.txt
crewai>=1.13.0
langchain>=0.1.0
pydantic>=2.0.0
openai>=1.0.0
flask>=3.1.0
flask-sqlalchemy>=3.1.0
werkzeug>=3.0.0
gunicorn>=21.0.0
sqlalchemy>=2.0.0
jira>=3.10.0
argon2-cffi>=21.0.0
pyjwt>=2.12.1
python-dotenv>=1.0.0
redis>=4.0.0
flask-talisman>=1.1.0
flask-wtf>=1.2.2
flask_login>=0.6.0
flask_limiter>=2.0.0
redis>=4.0.0
argon2-cffi>=21.0.0